Implement synchronous charging for a subscription
Due to how the FakePay gateway works, we have to synchronously
request a payment in order to get a credit card token, hence
the additional asynchronous step in `Transactions::CreateSubscription`
after the `persist!` step.

The FakePay gateway errors are also translated into a similar structure
as the validation errors, to make API more consistent.