Add transaction wrapper for operations
Due to the required synchronous nature of FakePay API card token
issuance, there is a need to only persist a new `Subscription` and
related models after the token is issued.

This could be ameliorated by separating the requirement for creating
the customer and subscription together, but didn't want to deviate too
much from the task description, which suggests a single endpoint.